+++ cmd/frontend/internal/graphqlbackend/search_repositories.go
@@ -13,17 +13,9 @@ var mockSearchRepositories func(args *repoSearchArgs) ([]*searchResultResolver, 
 //
 // For a repository to match a query, the repository's name ("URI") must match all of the repo: patterns AND the
 // default patterns (i.e., the patterns that are not prefixed with any search field).
-func searchRepositories(ctx context.Context, args *repoSearchArgs, query searchquery.Query) (res []*searchResultResolver, common *searchResultsCommon, err error) {
+func searchRepositories(ctx context.Context, args *repoSearchArgs, query searchquery.Query, limit int32) (res []*searchResultResolver, common *searchResultsCommon, err error) {
 	if mockSearchRepositories != nil {
 		return mockSearchRepositories(args)
-	}
-
-	// Only proceed if the query consists solely of type:, repo:, repogroup:, and default (no field) tokens.
-	// Matching repositories based whether they contain files at a certain path (etc.) is not yet implemented.
-	for field := range query.Fields {
-		if field != searchquery.FieldRepo && field != searchquery.FieldRepoGroup && field != searchquery.FieldType && field != searchquery.FieldDefault {
-			return nil, nil, nil
-		}
 	}
 
 	pattern, err := regexp.Compile(args.query.Pattern)
@@ -34,6 +26,10 @@ func searchRepositories(ctx context.Context, args *repoSearchArgs, query searchq
 	common = &searchResultsCommon{}
 	var results []*searchResultResolver
 	for _, repo := range args.repos {
+		if len(results) == int(limit) {
+			common.limitHit = true
+			break
+		}
 		if pattern.MatchString(string(repo.repo.URI)) {
 			results = append(results, &searchResultResolver{repo: &repositoryResolver{repo: repo.repo}})
 		}