+++ cmd/frontend/internal/app/pkg/updatecheck/handler.go
@@ -27,7 +27,7 @@ var (
 	// Server installations what the latest version is. The version here _must_ be
 	// available at https://hub.docker.com/r/sourcegraph/server/tags/ before
 	// landing in master.
-	latestReleaseServerBuild = newBuild("2.6.6")
+	latestReleaseServerBuild = newBuild("2.6.8")
 
 	// latestReleaseDataCenterBuild is only used by sourcegraph.com to tell existing
 	// Data Center installations what the latest version is. The version here _must_ be